+++ src/objects/cell.cpp
@@ -1,32 +1,33 @@
 #include "cell.hpp"
 
 namespace object {
-Cell::Cell(const std::vector<float> &points, const std::vector<float> &radii) {
+Cell::Cell(const std::vector<double> &points,
+           const std::vector<double> &radii) {
 
    if (points.size() != radii.size() * 3)
       throw std::invalid_argument("points and radii aren't same size");
 
    points_.resize(radii.size());
    points_.shrink_to_fit();
-   voi_ = aabb::AABB<float, 3>{};
+   voi_ = aabb::AABB<double, 3>{};
    radii_ = radii;
 
    for (size_t i = 0; i < radii_.size(); i++)
-      points_[i] = vm::Vec3<float>(points[3 * i + 0], points[3 * i + 1],
-                                   points[3 * i + 2]);
+      points_[i] = vm::Vec3<double>(points[3 * i + 0], points[3 * i + 1],
+                                    points[3 * i + 2]);
 
    // calc voi
    if (points_.empty())
       return;
    else if (points_.size() == 1) {
-      voi_ = aabb::AABB<float, 3>(points_[0], points_[0]);
+      voi_ = aabb::AABB<double, 3>(points_[0], points_[0]);
       return;
    }
    CalculateVoi();
 }
 
-Cell::Cell(const std::vector<vm::Vec3<float>> &points,
-           const std::vector<float> &radii) {
+Cell::Cell(const std::vector<vm::Vec3<double>> &points,
+           const std::vector<double> &radii) {
 
    if (points.size() != radii.size())
       throw std::invalid_argument("points and radii aren't same size");
@@ -38,7 +39,7 @@ Cell::Cell(const std::vector<vm::Vec3<float>> &points,
    if (points_.empty())
       return;
    else if (points_.size() == 1) {
-      voi_ = aabb::AABB<float, 3>(points_[0], points_[0]);
+      voi_ = aabb::AABB<double, 3>(points_[0], points_[0]);
       return;
    }
    CalculateVoi();
